refactor(grpc plugins) extract common grpc code
common functions for:

- loading a .proto file
  - from common subdirectories
  - installs "well-known types" transcoding (only Timestamp for now)
  - applies a function to analize each defined RPC method.
- (un)framing protobuf messages in a data stream.
