+++ kong-2.6.0-0.rockspec
@@ -131,6 +131,7 @@ build = {
     ["kong.status"] = "kong/status/init.lua",
 
     ["kong.tools.dns"] = "kong/tools/dns.lua",
+    ["kong.tools.grpc"] = "kong/tools/grpc.lua",
     ["kong.tools.utils"] = "kong/tools/utils.lua",
     ["kong.tools.timestamp"] = "kong/tools/timestamp.lua",
     ["kong.tools.stream_api"] = "kong/tools/stream_api.lua",
