+++ kong/plugins/grpc-web/deco.lua
@@ -1,15 +1,12 @@
 -- Copyright (c) Kong Inc. 2020
 
-require"lua_pack"
 local cjson = require "cjson"
-local protoc = require "protoc"
 local pb = require "pb"
-local pl_path = require "pl.path"
+local grpc_tools = require "kong.tools.grpc"
+local grpc_frame = grpc_tools.frame
+local grpc_unframe = grpc_tools.unframe
 
 local setmetatable = setmetatable
-
-local bpack=string.pack         -- luacheck: ignore string
-local bunpack=string.unpack     -- luacheck: ignore string
 
 local ngx = ngx
 local decode_base64 = ngx.decode_base64
@@ -62,26 +59,15 @@ local function get_proto_info(fname)
     return info
   end
 
-  local dir, name = pl_path.splitpath(pl_path.abspath(fname))
-  local p = protoc.new()
-  p.include_imports = true
-  p:addpath(dir)
-  local parsed = p:parsefile(name)
-
   info = {}
-
-  for _, srvc in ipairs(parsed.service) do
-    for _, mthd in ipairs(srvc.method) do
-      info[("/%s.%s/%s"):format(parsed.package, srvc.name, mthd.name)] = {
-        mthd.input_type,
-        mthd.output_type,
-      }
-    end
-  end
+  grpc_tools.each_method(fname, function(parsed, srvc, mthd)
+    info[("/%s.%s/%s"):format(parsed.package, srvc.name, mthd.name)] = {
+      mthd.input_type,
+      mthd.output_type,
+    }
+  end)
 
   _proto_info[fname] = info
-
-  p:loadfile(name)
   return info
 end
 
@@ -130,25 +116,6 @@ function deco.new(mimetype, path, protofile)
 end
 
 
-local function frame(ftype, msg)
-  return bpack("C>I", ftype, #msg) .. msg
-end
-
-local function unframe(body)
-  if not body or #body <= 5 then
-    return nil, body
-  end
-
-  local pos, ftype, sz = bunpack(body, "C>I")       -- luacheck: ignore ftype
-  local frame_end = pos + sz - 1
-  if frame_end > #body then
-    return nil, body
-  end
-
-  return body:sub(pos, frame_end), body:sub(frame_end + 1)
-end
-
-
 function deco:upstream(body)
   if self.text_encoding == "base64" then
     body = decode_base64(body)
@@ -157,10 +124,10 @@ function deco:upstream(body)
   if self.msg_encoding == "json" then
     local msg = body
     if self.framing == "grpc" then
-      msg = unframe(body)
+      msg = grpc_unframe(body)
     end
 
-    body = frame(0x0, pb.encode(self.input_type, decode_json(msg)))
+    body = grpc_frame(0x0, pb.encode(self.input_type, decode_json(msg)))
   end
 
   return body
@@ -172,17 +139,17 @@ function deco:downstream(chunk)
     local body = (self.downstream_body or "") .. chunk
 
     local out, n = {}, 1
-    local msg, body = unframe(body)
+    local msg, body = grpc_unframe(body)
 
     while msg do
       msg = encode_json(pb.decode(self.output_type, msg))
       if self.framing == "grpc" then
-        msg = frame(0x0, msg)
+        msg = grpc_frame(0x0, msg)
       end
 
       out[n] = msg
       n = n + 1
-      msg, body = unframe(body)
+      msg, body = grpc_unframe(body)
     end
 
     self.downstream_body = body
@@ -198,7 +165,7 @@ end
 
 
 function deco:frame(ftype, msg)
-  local f = frame(ftype, msg)
+  local f = grpc_frame(ftype, msg)
 
   if self.text_encoding == "base64" then
     f = ngx.encode_base64(f)
